-redefined contract for IMessagingService. Service now returns publisher status - added constants for publisher message statuses - refactored exceptions for publisher and consumer when messaging via a broker
