+++ CleanArchitecture.SharedLibrary/Messaging/Constants/PublisherMessageStatuses.cs
@@ -1,6 +1,6 @@
 ﻿namespace CleanArchitecture.SharedLibrary.Messaging.Constants
 {
-    public static class MessageServicePublishedMessageStatuses
+    public static class PublisherMessageStatuses
     {
         public const string Published = "published";
         public const string PossiblyPublished = "possibly-published";
